+++ CHANGELOG.md
@@ -8,6 +8,16 @@ If you like this project and find it useful, please consider giving it a star on
 
 <a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="120"></a>
 
+## [0.0.6] - 2026-06-10
+
+### Added
+
+- Require Matterbridge 3.9.0.
+- Add subscribe topic **matterbridge/deviceid/subscribe/root**.
+- Remove Jest and add Vitest.
+
+<a href="https://www.buymeacoffee.com/luligugithub"><img src="https://matterbridge.io/assets/bmc-button.svg" alt="Buy me a coffee" width="80"></a>
+
 ## [0.0.5] - 2026-06-09
 
 ### Added